+++ playwright-tests/e2e/8-settings/pmtConfiguration.spec.ts
